Added instance location tracking
The __current key will contain the name of the class the current function is running on. This is not tested and will likely need some adjustments before it's perfect.

Please note: The current property does NOT adjust itself after a function call like the super does. It is set before a function call and that is it.